requirements.txt
openai>=0.27.0
openai-whisper
